refactor: update README and enhance console plugin helpers with new merge options functionality
