+++ packages/logger-plugin-console/src/helpers/helper-merge-options.ts
